+++ CHANGELOG.md
@@ -5,6 +5,15 @@
 ### Added
 
   - Middleware - support content-encoding gzip when sent by http client
+  - System group `group/nuvla-vpn`
+
+### Changed
+
+  - Nuvlabox status - Update heartbeat even if NuvlaBox encounter a spec
+    validation issue
+  - Nuvlabox-status - Use delete attributes
+  - Job - Use delete-attributes
+  - Nuvlabox - Delete location possible
 
 ## [5.19.1] - 2021-11-10